right tridioagonal matrix made basic finite difference

+++ Koshi_giu/Cauchy_solver/Eiler_Cauchy.cs
@@ -10,28 +10,28 @@ namespace Cauchy
 {
     class Eiler_Cauchy :Cauchy_Solver
     {
-        public override double [ ] Solve( Func<double, double [ ], double> [ ] right_funcs, Value [ ] prev_value, double x )
+        public override decimal [ ] Solve( Func<decimal, decimal [ ], decimal> [ ] right_funcs, Value [ ] prev_value, decimal x )
         {
             dim = prev_value.Last().U.Length;
             h = x - prev_value.Last().X;
 
-            double [ ] res = Calculate_u(right_funcs,prev_value[0],x);
-            res = Correct( right_funcs,prev_value[0],new Value(x,res),0.00000001 );
+            decimal [ ] res = Calculate_u(right_funcs,prev_value[0],x);
+            res = Correct( right_funcs,prev_value[0],new Value(x,res),0.00000001M );
             return res;
         }
 
-        private double [ ] Calculate_u ( Func<double, double [ ], double> [ ] right_funcs, Value prev_value, double x )
+        private decimal [ ] Calculate_u ( Func<decimal, decimal [ ], decimal> [ ] right_funcs, Value prev_value, decimal x )
         {
                             
-            double [ ] res = new double [ dim ];
+            decimal [ ] res = new decimal [ dim ];
             for ( int p = 0 ; p < dim ; ++p )
                 res [ p ] = prev_value.U [ p ] + h * right_funcs [ p ] ( prev_value.X , prev_value.U );
             return res;
         }
 
-        private double[] Correct( Func<double, double [ ], double> [ ] right_funcs, Value prev_value, Value current_value, double eps )
+        private decimal[] Correct( Func<decimal, decimal [ ], decimal> [ ] right_funcs, Value prev_value, Value current_value, decimal eps )
         {
-            double [ ] next_iteration = new double [ dim ];
+            decimal [ ] next_iteration = new decimal [ dim ];
             current_value. U . CopyTo ( next_iteration , 0 );
             do
             {
